+++ UI/InputINUI.cs
@@ -70,7 +70,7 @@ namespace MyHome2013
                 var ich = Program.Container.GetInstance<IRepository<IncomeCategory>>();
 
                 Income newIncome = 
-                    new Income(double.Parse(this.txtAmount.Text), this.dtPick.Value,
+                    new Income(decimal.Parse(this.txtAmount.Text), this.dtPick.Value,
                                 ich.LoadById(Convert.ToInt32(this.cmbCategory.SelectedValue)),
                                 r.LoadById(Convert.ToInt32(this.cmbPayment.SelectedValue)),
                                 this.txtDetail.Text);
